linux port

+++ src/ColumnString.cpp
@@ -1,8 +1,9 @@
+#include <cstdlib>
+#include <cassert>
+#include <cstring>
+#include <cstdio> // debug
+
 #include "Column.h"
-#include <stdlib.h>
-#include <assert.h>
-#include <string.h>
-#include <stdio.h> // debug
 
 AdaptiveStringColumn::AdaptiveStringColumn() : Column(COLUMN_NORMAL) {
 }
@@ -199,7 +200,7 @@ bool AdaptiveStringColumn::Alloc(size_t count, size_t width) {
 		m_capacity = new_capacity;
 
 		// Update ref in parent
-		UpdateParent((int)data);
+		UpdateParent((uintptr_t)data);
 	}
 
 	// Pack width in 3 bits (log2)
@@ -233,8 +234,8 @@ size_t AdaptiveStringColumn::Find(const char* value) const {
 size_t AdaptiveStringColumn::Find(const char* value, size_t len) const {
 	assert(value);
 
-	if (m_len == 0) return (size_t)-1; // empty list
-	if (len >= m_width) return (size_t)-1; // A string can never be wider than the column width
+	if (m_len == 0) return -1; // empty list
+	if (len >= m_width) return -1; // A string can never be wider than the column width
 
 	if (m_width == 0) {
 		return 0; 
@@ -290,7 +291,7 @@ size_t AdaptiveStringColumn::Find(const char* value, size_t len) const {
 	}
 	else assert(false);
 		
-	return (size_t)-1;
+	return -1;
 }
 
 void AdaptiveStringColumn::Stats() const {
